requirements.txt
fastapi~=0.115.12
matplotlib
torch
numpy
python-dotenv~=1.1.0
uvicorn
google-api-python-client~=2.171.0
python-multipart
pydantic~=2.11.5
textblob
starlette